[Step 2] 리뷰 반영
- 숫자 체크 정규식 제거
- 사칙연산자 Enum 화, 연산 로직도 Enum의 함수로 구현
- null이 될 수 없는 인자들 ? 제거
- null 테스트 제거
