+++ src/main/kotlin/stringcalculator/application/StringCalculator.kt
@@ -7,7 +7,7 @@ import stringcalculator.util.StringValidator
 class StringCalculator(
     private val operatorSelector: OperatorSelector
 ) {
-    fun calculate(string: String?): Double {
+    fun calculate(string: String): Double {
         StringValidator.validateNotBlank(string)
 
         val trimmedValue = string!!.removeRepeatedBlank()
@@ -25,15 +25,12 @@ class StringCalculator(
         return result
     }
 
-    companion object ArgumentsValidator {
-
-        fun validateLength(split: List<String>) {
-            if (split.size < 3) {
-                throw IllegalArgumentException("계산할 숫자와 연산자가 충분하지 않습니다.")
-            }
-            if (split.size % 2 == 0) {
-                throw IllegalArgumentException("숫자 또는 연산자가 누락되었습니다.")
-            }
+    private fun validateLength(split: List<String>) {
+        if (split.size < 3) {
+            throw IllegalArgumentException("계산할 숫자와 연산자가 충분하지 않습니다.")
+        }
+        if (split.size % 2 == 0) {
+            throw IllegalArgumentException("숫자 또는 연산자가 누락되었습니다.")
         }
     }
 }
