--- src/main/kotlin/stringcalculator/operator/MultiplyOperator.kt
@@ -1,11 +1,0 @@
-package stringcalculator.operator
-
-class MultiplyOperator : Operator {
-    override fun character(): String {
-        return "*"
-    }
-
-    override fun operate(x: Double, y: Double): Double {
-        return x * y
-    }
-}
