--- src/main/kotlin/stringcalculator/operator/OperatorRegister.kt
@@ -1,10 +1,0 @@
-package stringcalculator.operator
-
-object OperatorRegister {
-    val operators = listOf(
-        PlusOperator(),
-        MinusOperator(),
-        MultiplyOperator(),
-        DivideOperator()
-    )
-}
